Game Design, Map1 repair, Detail map fix
git-tfs-id: [http://192.168.56.101:8080/tfs/DefaultCollection]$/TikiEngine 2.0;C1272

+++ misc/tools/TikiEditor/Source/TikiEditor.Cloddy/TikiEditor.Cloddy.cpp
@@ -69,7 +69,7 @@ namespace TikiEditor
 				&sample
 			);
 
-			data[i] = sample.Elevation;
+			data[(int)i] = sample.Elevation;
 
 			i++;
 		}
